Try to auto-bind to the partner's domid
By specifying a partner id of XEN_ARGO_DOMID_ANY, we are creating a
wildcard ring.  This is sub-optimal since it requires a broader XSM
policy than a domain specific single-source ring.

Since we are connecting, we usually have a connect_addr with a partner
domid.  Specify that domid as the partner_id for the bind operation
which will create a single source ring.

(bind can also take a NULL connect_addr argument to detach a socket.  In
that case, the socket state should not be ARGO_STATE_IDLE.  This may
make sense for re-assigning a previously used argo ring.)

auto-bind is primarily for when a socket is created and then immediately
connect is called - a typical client socket connection.  Therefore it
seems acceptable to make the above behaviour change.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>